More support for older boringssl+fips
